+++ assets/javascripts/kitten/components/action/button-image/index.js
@@ -4,7 +4,6 @@ import classNames from 'classnames'
 import styled from 'styled-components'
 import { pxToRem } from '../../../helpers/utils/typography'
 import { ScreenConfig } from '../../../constants/screen-config'
-import { checkDeprecatedSizes } from '../../../helpers/utils/deprecated'
 
 const StyledButtonImage = styled.button`
   width: ${pxToRem(40)};
@@ -25,13 +24,11 @@ const StyledButtonImage = styled.button`
     opacity: 0.8;
   }
 
-  &.k-ButtonImage--tiny,
   &.k-ButtonImage--small {
     width: ${pxToRem(30)};
     height: ${pxToRem(30)};
   }
 
-  &.k-ButtonImage--big,
   &.k-ButtonImage--large {
     @media (min-width: ${ScreenConfig.S.min}px) {
       width: ${pxToRem(50)};
@@ -72,7 +69,6 @@ export const ButtonImage = ({
   img,
   ...others
 }) => {
-  checkDeprecatedSizes(size)
 
   return (
     <StyledButtonImage