+++ assets/javascripts/kitten/components/form/label/index.js
@@ -6,7 +6,6 @@ import isEmpty from 'lodash/fp/isEmpty'
 import { domElementHelper } from '../../../helpers/dom/element-helper'
 import TYPOGRAPHY from '../../../constants/typography-config'
 import { pxToRem, stepToRem } from '../../../helpers/utils/typography'
-import { checkDeprecatedSizes } from '../../../helpers/utils/deprecated'
 
 const StyledLabel = styled.label`
   display: block;
@@ -53,7 +52,6 @@ export const Label = ({
   style,
   ...other
 }) => {
-  checkDeprecatedSizes(size)
 
   const handleClick = e => {
     if (domElementHelper.canUseDom() && focusId) {