+++ assets/javascripts/kitten/components/form/select/dropdown-select/combobox.js
@@ -7,7 +7,6 @@ import classNames from 'classnames'
 import { WarningCircleIcon } from '../../../graphics/icons/warning-circle-icon'
 import { CheckedCircleIcon } from '../../../graphics/icons/checked-circle-icon'
 import { ArrowIcon } from '../../../graphics/icons/arrow-icon'
-import { checkDeprecatedSizes } from '../../../../helpers/utils/deprecated'
 import find from 'lodash/fp/find'
 import flow from 'lodash/fp/flow'
 import uniqBy from 'lodash/fp/uniqBy'
@@ -57,7 +56,6 @@ export const DropdownCombobox = ({
   labelProps,
   inputProps,
 }) => {
-  checkDeprecatedSizes(size)
   const [flattenedOptions, setFlattenedOptions] = useState([])
   const [filteredOptions, setFilteredOptions] = useState([])