+++ assets/javascripts/kitten/components/graphics/illustrations/kkbb-flash-illustration.js
@@ -1,9 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { checkDeprecatedSizes } from '../../../helpers/utils/deprecated'
 
 export const KkbbFlashIllustration = ({ size, ...props }) => {
-  checkDeprecatedSizes(size)
 
   if (size === 'small') {
     return (