+++ assets/javascripts/kitten/components/typography/horizontal-stroke/index.js
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
 import COLORS from '../../../constants/colors-config'
 import { pxToRem } from '../../../helpers/utils/typography'
-import { checkDeprecatedSizes } from '../../../helpers/utils/deprecated'
 import classNames from 'classnames'
 import { strokeModifierStylesWithoutMargin } from './common/stroke-modifier-styles'
 
@@ -27,10 +26,6 @@ export const LARGE = css`
   height: ${pxToRem(4)};
 `
 
-/* DEPRECATED EXPORTS */
-export const TINY = SMALL
-export const DEFAULT = MEDIUM
-export const BIG = LARGE
 export const HUGE = css`
   width: ${pxToRem(100)};
   height: ${pxToRem(6)};
@@ -44,20 +39,14 @@ export const StyledHorizontalStroke = styled.div`
     ${MICRO}
   }
 
-  &.k-HorizontalStroke--size--tiny {
-    ${TINY}
-  }
-  &.k-HorizontalStroke--size--tiny, /* DEPRECATED */
   &.k-HorizontalStroke--size--small {
     ${SMALL}
   }
 
-  &.k-HorizontalStroke--size--default, /* DEPRECATED */
   &.k-HorizontalStroke--size--medium {
     ${MEDIUM}
   }
 
-  &.k-HorizontalStroke--size--big, /* DEPRECATED */
   &.k-HorizontalStroke--size--large {
     ${LARGE}
   }
@@ -78,7 +67,6 @@ export const HorizontalStroke = ({
   color,
   ...props
 }) => {
-  checkDeprecatedSizes(size)
 
   const modifierClassName = () => {
     if (!modifier) return `k-HorizontalStroke--size--${size}`